metrics: add multiRegistry type to include k8s ctrl runtime metrics.
The K8s controller runtime library registers a lot of its own metrics
to the prometheus internal registry.

This means that when switching the Operator to the hive provided metrics
registry we will be missing controller-runtime metrics.

To prevent breaking existing metrics, as surfaced by the `metrics list`
command or via the prometheus http api server, we create a multiRegistry
type that implements the prometheus.Gatherer interface for a set of
registries.  With this we include the global registry when doing
`metrics list` and as the registry gatherer passed to the prometheus
http server.

As a drawback, this means that the metrics sampler will not have access
to controller runtime metrics. But this may require upstream changes in
controller-runtime to fix reliably, while at the same time we seek to
not break existing metrics functionality.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>