plugins: Don't install CNI conf in container image
I noticed in local testing that when you deploy Cilium with a custom CNI
configuration, Cilium may report a different CNI configuration file
inside the Cilium container compared with the configuration on the
underlying host despite mapping the host filesystem into the Cilium
container. This appears to be due to the use of overlayfs and because
the 'install-binary' target (invoked during container build) installs
this config file into the filesystem of the container itself. This
ultimately obscures the true config when inspecting it from inside the
Cilium container.

Move the binary install steps into the 'install-binary' target, and
reverse the dependency so that if someone wishes to install the CNI from
the tree they can invoke the 'install' target to install both the binary
and this default config.

Signed-off-by: Joe Stringer <joe@cilium.io>